+++ R/utils-faces.R
@@ -116,3 +116,89 @@ top_y <- function(mat, angle, scale, width) {
         row[2] + y_vertices + z_factor * (row[3] + 0.5 * width)
     }))
 }
+
+face_x_oid <- function(face, mat, angle, scale, width) {
+    switch(face,
+           top = top_x_oid(mat, angle, scale, width),
+           east = east_x_oid(mat, angle, scale, width),
+           west = west_x_oid(mat, angle, scale, width),
+           north = north_x_oid(mat, angle, scale, width),
+           south = south_x_oid(mat, angle, scale, width)
+    )
+}
+face_y_oid <- function(face, mat, angle, scale, width) {
+    switch(face,
+           top = top_y_oid(mat, angle, scale, width),
+           east = east_y_oid(mat, angle, scale, width),
+           west = west_y_oid(mat, angle, scale, width),
+           north = north_y_oid(mat, angle, scale, width),
+           south = south_y_oid(mat, angle, scale, width)
+    )
+}
+
+north_x_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_x(scale, angle)
+    x_vertices <- -width * c(0.5, -0.5, -0.5, 0.5)
+    as.numeric(apply(mat, 1, function(row) {
+        row[1] + x_vertices + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+north_y_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_y(scale, angle)
+    as.numeric(apply(mat, 1, function(row) {
+        row[2] + 0.5 * width + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+east_x_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_x(scale, angle)
+    as.numeric(apply(mat, 1, function(row) {
+        row[1] + 0.5 * width + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+east_y_oid <-  function(mat, angle, scale, width) {
+    z_factor <- z_factor_y(scale, angle)
+    y_vertices <- width * c(0.5, -0.5, -0.5, 0.5)
+    as.numeric(apply(mat, 1, function(row) {
+        row[2] + y_vertices + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+south_x_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_x(scale, angle)
+    x_vertices <- -width * c(0.5, -0.5, -0.5, 0.5)
+    as.numeric(apply(mat, 1, function(row) {
+        row[1] + x_vertices + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+south_y_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_y(scale, angle)
+    as.numeric(apply(mat, 1, function(row) {
+        row[2] - 0.5 * width + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+west_x_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_x(scale, angle)
+    as.numeric(apply(mat, 1, function(row) {
+        row[1] - 0.5 * width + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+west_y_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_y(scale, angle)
+    y_vertices <- width * c(0.5, -0.5, -0.5, 0.5)
+    as.numeric(apply(mat, 1, function(row) {
+        row[2] + y_vertices + z_factor * c(0, 0, row[3], row[3])
+    }))
+}
+top_x_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_x(scale, angle)
+    x_vertices <- width * c(-0.5, -0.5, 0.5, 0.5)
+    as.numeric(apply(mat, 1, function(row) {
+        row[1] + x_vertices + z_factor * (row[3])
+    }))
+}
+top_y_oid <- function(mat, angle, scale, width) {
+    z_factor <- z_factor_y(scale, angle)
+    y_vertices <- width * c(0.5, -0.5, -0.5, 0.5)
+    as.numeric(apply(mat, 1, function(row) {
+        row[2] + y_vertices + z_factor * (row[3])
+    }))
+}