+++ R/utils-transform.R
@@ -1,13 +1,13 @@
 op_sort <- function(df, angle=45) {
     angle <- angle %% 360
     if ((0 <= angle) && (angle < 90)) {
-        df <- df[order(df$z, -df$y, -df$x), ]
+        df <- df[order(-df$y, -df$x, df$z), ]
     } else if ((90 <= angle) && (angle < 180)) {
-        df <- df[order(df$z, -df$y, df$x), ]
+        df <- df[order(-df$y, df$x, df$z), ]
     } else if ((180 <= angle) && (angle < 270)) {
-        df <- df[order(df$z, df$y, df$x), ]
+        df <- df[order(df$y, df$x, df$z), ]
     } else {
-        df <- df[order(df$z, df$y, -df$x), ]
+        df <- df[order(df$y, -df$x, df$z), ]
     }
     df
 }